requirements.txt
Django==2.2.1
djangorestframework
djangorestframework-jwt
cryptography
pyjwt
python-dotenv


